requirements.txt
gunicorn
uvicorn
fastapi == 0.115.4
psycopg == 3.1.19
sqlalchemy
pydantic_settings
pydantic[email]
alembic
psycopg-binary ==3.1.19
bcrypt==3.2.2
passlib
pyjwt == 2.8.0
jinja2
starlette ==0.41.2
itsdangerous == 2.2.0
pyi18n-v2 == 1.2.2

